backend/requirements.txt
flask
flask-cors
werkzeug
google-cloud-documentai
pypdf
openai
python-dotenv
